% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iopolicy-package.R
\docType{package}
\name{iopolicy-package}
\alias{iopolicy}
\alias{iopolicy-package}
\title{iopolicy: shared infrastructure for industrial-organization policy models}
\description{
The package provides model-independent synthetic market design and
reproducibility helpers. Economic equations and model-specific realizations
remain in downstream packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luciu5/iopolicy}
}

}
\author{
\strong{Maintainer}: Charles Taragin \email{ctaragin+antitrustr@gmail.com}

Authors:
\itemize{
  \item Charles Taragin \email{ctaragin+antitrustr@gmail.com}
}

}
\keyword{internal}
