% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntheticMarket.R
\name{market_design}
\alias{market_design}
\title{Return synthetic market design metadata}
\usage{
market_design(x)
}
\arguments{
\item{x}{A `SyntheticMarket` object.}
}
\value{
The market's design metadata.
}
\description{
Return synthetic market design metadata
}
