% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntheticMarket.R
\name{print.SyntheticMarket}
\alias{print.SyntheticMarket}
\title{Print a synthetic market}
\usage{
\method{print}{SyntheticMarket}(x, ...)
}
\arguments{
\item{x}{A `SyntheticMarket` object.}

\item{...}{Ignored.}
}
\description{
Print a synthetic market
}
